还差motor lib

+++ source/motor_control/Core/Src/pid.c
@@ -1,4 +1,4 @@
-float pid_init(int kp,int ki,int kd)
+float pid_init(float kp,float ki,float kd)
 {
     return 0;
 }
